requirements.txt
spotipy
python-dotenv
jellyfish
flask